cldr2def: Improve ctype makefile generation to avoid symlinks
I was informed that symlinks in repository might cause problems
when the repository is extracted to windows.  The generated ctype
makefile was improved to work without symlinks at all.  (The symlinks
were there to avoid multiple copies of the same exact file, some of
which were greater than 1 Mb in size).