kernel: Handle zero-length ELF sections better
FreeBSD simply skips all zero-length ELF sections when scanning an
REL type file.  The first time we tried it, DF panicked with a lost
bigs message.  It turns out this was because the skip should have been
applied in 3 places and we missed the first one.

The same approach is needed by the boot loader as well.