<inttypes.h>: Expose wcstoimax() and wcstoumax() prototypes for C++ too.
I don't know why it was explicitly excluded in e2565a420c31.

I can only guess that because of the definition of wchar_t being in
"#ifndef __cplusplus", it was assumed that these functions (both
taking wchar_t's) should get the same treatment. However, the real
reason for the wchar_t handling is that wchar_t is a fundamental
type in C++, while in C it is not.