libc: Remove unused (and public) locale_isstrict()
The publicly exposed locale_isstrict function is not hooked into anything.
It was intended to implement LOCALE_STRICT environment variable handling,
but it only returns a boolean based on the value of LOCALE_STRICT, and not
even efficiently.

There is also a private "strict_mode" variable in locale_private.h that
apparently was intended to be used with LOCALE_STRICT, but it's also
seemingly useless.

The LOCALE_STRICT override still needs to be implemented if UTF-8B is
in use, but it can be handled at encoding initialization more
efficiently later.