<stdlib.h>/<unistd.h>: Expose mkdtemp()'s prototype also if XSI is visible.
The history of this function in the specs seems to be a bit convoluted
for this function, to say the least. Even though I can't find it in
either Issue 5 or 6, go with what Issue 7's HISTORY says and treat it
like it was moved from XSI to base starting with Issue 7:

http://pubs.opengroup.org/onlinepubs/9699919799/functions/mkdtemp.html

Most importantly, this commit fixes x11/xterm which was building but
segfaulting after 8eb95a5667325.

Reported-by: Rimvydas Jasinskas <rimvydas.jasinskas@gmail.com>