sys/dev/disk/dm: Make target's init() take argc and argv
dm core shouldn't be passing char *params to targets' init()
because all targets are basically going to copy-paste a code
snippet (that strseps params) from the existing targets code
which is inefficient and should be avoided.

This commit adds dm_table_init() which converts char *params
to int argc and char **argv, and passes them to init().