sys/dev/disk/dm: Fix sanity checks for striped target [4/4]
Add sanity checks on relation among
number of stripes, device size, and chunk size.

This is one of the requirements in Linux dm-striped, but
disable it for now as it may break systems that currently
use this target (which I doubt there is any because lvm2
and dm raid seem to be unusable on DragonFly).