libelf: Add libelf to lib directory, but leave it unhooked
The ELF library is part of FreeBSD's ELF Tool Chain project that has the
goal to be a drop-in replacement for binutils.  It isn't quite there yet
but the toolchain project has replaced most of the tools.  There may be
interest to bring the entire toolchain into DragonFly once the project is
production ready, but for now we are only interested in libelf.

This library is a required component for the incoming execinfo library.
The ELF library is for internal use only.  It will not be installed, but
instead only statically linked to the execinfo library.  The
contrib/elftoolchain directory has been arranged in a "placeholder"
fashion so the source for more of the tools can be added there later.