Add libexecinfo to lib directory, but leave it unhooked
The execinfo library is used to fill in the backtrace of the currently
executing thread.  It originated in Linux and has found its way into a
lot of vendor software.  NetBSD created a BSD-licensed implementation
which FreeBSD imported a while back.

The execinfo library has been in FreeBSD since Release 10.  FreeBSD 9
is the only release that needs libexecinfo from ports and when that
release hits EOL, DragonFly could be left unsupported if ports
maintainers assumes libexecinfo is always available and removes USES
support for it.

To avoid this situation, bring libexecinfo into DragonFly's base for
the purpose of supporting ports.  It requires libelf which has also
been brought in as an internal library.