lib/: Fix library build dependencies
Several libraries were not listed in the SUBDIR_ORDERED definition,
but should have been.  Many of these were accounted for in the
_prebuild_libs functionality of Makefile.inc1.  The rest were far
down on the alphabetical list and probably satisfied their dependency
requirements by luck (implying a machine with a sufficiently high
-j number would fail so lib/ wasn't jobs safe).

I've listed all the libraries that link to other libraries or have
other libraries (within lib/) link to them in SUBDIR_ORDERED.  I've
completed the requirement annotations and separated them
requirements that are handled by Makefile.inc1 so it's clear. This
results with seven libraries moving from the alphabetical list to
the ordered list.

This makes the "serial" part of the lib/ building pretty long and
unnecessarily long, but it's not possible to get to a lower granularity
with the current framework without creating a new level (having the
subdirectories with their own library subdirectories, similar to
how gcc50 was implemented).