contrib/dialog: Delete entire contents
There is a vendor/DIALOG branch which is supposed to feed in the
contents of the contrib/dialog directory via a vendor branch merge.
It was never merged though.  The contents of vendor/DIALOG are just
a copy of the vendor/branch so it's impossible to update via the
normal update-vendor-branch-merge-again process.

To fix, we delete everything and merge for the first time.  The
next commit will merge the dialog source updates.