locales: Remove two more aliases (ja_JP.eucjp and en_US.ISO-8859-1)
There was an earlier effort to define common locales (arising from
Linux's case-and-hyphen insensitivity) and they were going to be
removed when our locale handling followed Linux behavior, but it was
decided that we would and should not follow Linux's example.

Since then and through collaborative discussions with FreeBSD, it
was decided these common-but-incorrect-for-BSD aliases should not
continue.  FreeBSD will never get them and I'm removing ours.