compilers.conf: Add the two latest clangs to options
This adds "clang37" and "nextclang" defaults to compilers.conf.
Note that nextclang has a new port: devel/llvm-devel.  In fact
lang/clang37 is just a metaport that builds devel/llvm37.

"nextclang" will change over time.  Currently it's the precursor to
clang38, but once llvm3.8 is released, it will become the precursor
to clang 3.9 (or whatever comes after 3.8)