sys/kern/vfs: Fix wrong function name in kprintf/panic
ae8e83e6e in 2009 renamed biodone() to bpdone(), and made
bpdone() to be called by the new biodone() implementation
as well as other callers, but the function names used in
kprintf() and panic() were not updated to "bpdone:".

bpdone() isn't exclusively used by biodone(), so it should
use a proper function name.