acpi_pvpanic: Notify Qemu VM host if we panic.
By default the virtual machine is stopped when the pvpanic is triggered.
For accessing ddb, the virtual machine execution then needs to be resumed
manually (e.g. with the "cont" command on the Qemu-monitor console).

Triggering a pvpanic instead of just continuing to the ddb prompt can be
useful to avoid hogging the CPU while hanging in ddb, and to provide a
kernel panic notification for the VM-host system.