sys/dev/disk/dm: Don't expect destroy() to set config to NULL
Having a comment like below in dm core that expects programmers
to code their dm targets in a certain way is not efficient.
They may not read such a rule written in the source.

  /*
   * Remove target specific config data. After successfull
   * call table_en->target_config must be set to NULL.
   */

Just check a return value of destroy() and set config to NULL
on dm core side since it's also visible to dm core. There is
nothing targets can/would do with the config pointer other than
setting it to NULL considering targets have just freed it.