sys/vfs/hammer: Remove signature from ondisk node
The node signature was added in the early stage of hammer
development, but it's no longer used for anything except
that newfs_hammer sets the field to GOOD (for no reason).

c461935c added some code to propagate the GOOD signature
on node split (just to make it look a bit more sane), but
it's useless anyway.

This commit changes the signature field to reserved00,
and also gets rid of HAMMER_BTREE_SIGNATURE_XXX macros.
The change doesn't affect anything just like c461935c did
nothing good or bad.