i386 removal, part 46/x: Remove sys/cpu/i386.
Move the four i386 headers which libstand previously took from this
directory to libstand's dir. This is just a precaution and actually
not necessary at all because our x86_64 headers have #ifdef __i386__
checks for everything which libstand needs (binary comparison with
a libstand that was compiled with only x86_64 headers confirms it).

Review of the .depend files in /usr/obj confirms that this headers
are now taken.

Also reduce some no longer needed compatibility setup in libstand's
Makefile.