boot0 refactoring
The patch provides refactoring for the sys/boot/pc32/boot0.S
assembly file to imrove readability and comprehensibility of
assembly code.

The main changes are:

Two new symbols added: the _SECTOR_FIELD_OFF which provides
offset for the 'sector' field in CHS and the _DATA_OFF which
provides offset for the boot0 data from the partition table.
We can use these symbols instead of magic numbers in the code.

The second change is updating comments in some place.

The patch was tested as in virtual machine as in real hardware.