sbin/fdisk: cleanups
This patch provides set of cleanups for the sbin/fdisk util.

1. MBRSIGOFF macro removed. Use the DOSMAGICOFFSET macro instead
which is already defined in the <sys/diskmbr.h>.

2. Global variable iotest removed. It was used in the write_sr0()
function before. We checked this variable and it it was set we
just printed information about zero sector and exited from the
write_sr0(). Actually, this variable wasn't set never and never
used anywhere. For this point we already have the `-t` flag
which prevents writing of sector 0.

3. Unused macros removed.

4. Unused functions removed.

5. MAX_SECTORS_PER_TRACK and MIN_SECTORS_PER_TRACK macros are added.

6. Fixed order of DIOCWLABEL ioctl in the write_sr0().

7. Indentation fixed.