privatize libedit (used by numerous base programs)
Besides affecting far more programs than the first two conversions,
this privatization requires a modification to bin/sh which is a bootstrap
program.  If bootstrapping, it will use libedit.* if it exists, otherwise
it will use libprivate_edit.*.  After "make upgrade", /usr/lib/libedit will
no longer be on the system so the private version will then be used in all
cases.