privatize libncurses (touches many, many programs)
The ncurses software produces 8 libraries: normal and wide versions of
ncurses, forms, menu, and panel.  The dragonfly base doesn't use forms
or menu at all so we can remove 4 libraries immediately and their
associated man pages.

The panel library is only used in one place, so technically libpanelw
is no longer required, but let's keep building it anyway.

While here, delete references to libcurses.  It was just symlinked to
ncurses, so let's take the opportunity to only reference ncurses from
this point forward and not install a private version of libcurses.