Makefile_upgrade.inc: Expand all wildcards from the lib privatization.
We want to see exactly what we remove, not just approximately.
'make upgrade' changes are a pain to review when everything is
muddled with patterns. And it gets worse over time because
without a not yet upgraded system one has no choice but to
restore a world from the time and replaying the 'make upgrade'
to see what was going on.

Also, if we ever need to bring back one file which is covered
by a pattern later on (and we've had such cases), no one will
know what the rest of the wildcard was covering aside from
that file.