Remove FreeBSD OS/ABI from raid driver blobs.
This fixes kgdb issues when the kernel was linked with ld.gold.

The problem was caused by ld.gold's behaviour of setting the OS/ABI in
the output binary to FreeBSD when any of the input objects is tagged
with the FreeBSD OS/ABI.

The hpt27xx, hptmv and hptrr raid-drivers are using binary blobs, with
the FreeBSD OS/ABI flag, so we should just set the OS/ABI of these
objects to none after uudecode-ing them.