libedit: Restore installation of the editrc(5) manual page.
Removal of manual pages for the library privatization affects only
section 3 manual pages (for the functions in the library which is
now private). editrc.5 describes the ~/.editrc file that has an
effect on all things linked against the privatized edit library
too, so we want to ship documentation for it.