sys/vfs/hammer: Rename struct hammer_entry_data to hammer_direntry_data
hammer_entry_data is a structure for dirctory entry, so it's
simply more understandable if the name was hammer_direntry_data.
(Note that rectype for directory entry is HAMMER_RECTYPE_DIRENTRY)

This change does not affect userspace since userspace applications
never use (and should never use) this structure, except for some
sbin/hammer commands. sbin/hammer/cmd_recover.c is the only file
that refers to the name of this structure.