ncurses: Upgrade version 5.9 (20110402) => 6.0 (20160305)
The tic program has been updated at the same time.
The symbol maps were expanded using the vendor symbol maps so they
should be more accurate than before.  Only one new symbol came in
for the 6.0 version, but I didn't see it present in any library.

For now, the ncurses headers are still being installed.  They may
be removed later if dports will support it and it doesn't cause
problem with unix specifications.  (On DragonFly, the ncurses
base libraries are private so in theory the headers should be too).